+++ src/rtlib/file_winputstr.c
@@ -23,10 +23,10 @@ FBCALL FB_WCHAR *fb_FileWstrInput( ssize_t chars, int fnum )
     dst = fb_wstr_AllocTemp( chars );
     if( dst != NULL )
     {
-        size_t read_chars = 0;
+        ssize_t read_chars = 0;
         if( FB_HANDLE_IS_SCREEN(handle) )
         {
-            while( read_chars != (size_t)chars )
+            while( read_chars != chars )
             {
                 res = fb_FileGetDataEx( handle,
                                         0,
